pd_picpurify/requirements.txt
fastapi
sniffio
uvicorn[standard]
SQLAlchemy
pyodbc
python-decouple
minio
aiofiles
aiohttp